Flattened UI (Hope you like it)
Created LoadingPage. All that's left is to find a high quality animated gif. If you find one, make sure it's named loading.gif
Or you can use the animation from the SearchProgressPageV2 and just remove the code for an image.

For CreatePlaylistPage I moved the Group name to the Table Column. So when you add a group name, it's on the table column. It's a bit cleaner like that.

Cleaned up some CSS.